#9 Move ID class (specs → typing)
